Update Scene Flow (#86)
* change title scene vite

* fix scene and bgm bugs

+++ space_invader/src/scene/TitleScene.ts
@@ -65,7 +65,7 @@ export default class TitleScene extends Phaser.Scene {
 		])
 	}
 
-	async create() {
+	create() {
 		const {width,height} = this.scene.scene.scale
 		const queryString = window.location.search
 		const urlParams = new URLSearchParams(queryString)
@@ -77,23 +77,6 @@ export default class TitleScene extends Phaser.Scene {
 		// 	this.scene.launch('setup')
 		// 	return
 		// }
-
-		const auth = getAuth();
-		auth.useDeviceLanguage();
-		(async ()=> {
-			await setPersistence(auth, browserSessionPersistence)
-			const user = auth.currentUser
-
-			if (user === null) {
-				this.scene.pause()
-				this.scene.launch('start-login')
-				return
-			}
-
-			// TODO check user data
-			// this.scene.pause()
-			// this.scene.launch('register')
-		})()
 
 		this.add
 			.tileSprite(0, 0, width, height, 'titleBackground')
@@ -139,7 +122,7 @@ export default class TitleScene extends Phaser.Scene {
 		}
 	}
 
-	update() {
+	async update() {
 		if (
 			this.hasController &&
 			(this.controller1?.direction.LEFT ||
@@ -147,16 +130,42 @@ export default class TitleScene extends Phaser.Scene {
 				this.controller1?.buttons.B7 > 0 ||
 				this.input.pointer1.isDown)
 		) {
-			this.startGame()
+			await this.startGame()
 		}
 	}
 
-	startGame() {
-		I18nSingleton.getInstance().destroyEmitter()
+	async startGame() {
+		if(import.meta.env.VITE_START_SCEN) {
+			// testing flow
+			this.scene.start(import.meta.env.VITE_START_SCENE || 'home', {
+				bgm: this.bgm,
+			})
+		}
+		else {
+			// normal flow
+			const auth = getAuth();
+			auth.useDeviceLanguage();
+			(async ()=> {
+				await setPersistence(auth, browserSessionPersistence)
+				const user = auth.currentUser
+			
+				if (user === null) {
+					this.scene.stop()
+					this.scene.launch('start-login', {bgm: this.bgm,})
+					return
+				}
+				else {
+					this.scene.stop()
+					this.scene.start('home', {bgm: this.bgm,})
+				}
 
-		this.scene.start(import.meta.env.VITE_START_SCEN || 'home', {
-			bgm: this.bgm,
-		})
+					// TODO check user data
+					// this.scene.pause()
+					// this.scene.launch('register')
+			})()
+			I18nSingleton.getInstance().destroyEmitter()
+		}
+
 		//this.scene.start(import.meta.env.VITE_START_SCEN || 'setting')
 		// import.meta.env.VITE_START_SCENE && new SoundManager(this).stop(this.bgm!)
 	}
